+++ packages/dd-trace/src/debugger/devtools_client/index.js
@@ -5,7 +5,7 @@ const { breakpoints } = require('./state')
 const session = require('./session')
 const { getLocalStateForCallFrame } = require('./snapshot')
 const send = require('./send')
-const { getScriptUrlFromId } = require('./state')
+const { getStackFromCallFrames } = require('./state')
 const { ackEmitting, ackError } = require('./status')
 const { parentThreadId } = require('./config')
 const log = require('../../log')
@@ -66,16 +66,7 @@ session.on('Debugger.paused', async ({ params }) => {
     thread_name: threadName
   }
 
-  const stack = params.callFrames.map((frame) => {
-    let fileName = getScriptUrlFromId(frame.location.scriptId)
-    if (fileName.startsWith('file://')) fileName = fileName.substr(7) // TODO: This might not be required
-    return {
-      fileName,
-      function: frame.functionName,
-      lineNumber: frame.location.lineNumber + 1, // Beware! lineNumber is zero-indexed
-      columnNumber: frame.location.columnNumber + 1 // Beware! columnNumber is zero-indexed
-    }
-  })
+  const stack = getStackFromCallFrames(params.callFrames)
 
   // TODO: Send multiple probes in one HTTP request as an array (DEBUG-2848)
   for (const probe of probes) {
